+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -22,8 +22,11 @@
  */
 namespace MeCms\Test\TestCase\Model\Table;
 
+use ArrayObject;
 use Cake\Cache\Cache;
+use Cake\Event\Event;
 use Cake\I18n\Time;
+use Cake\ORM\Entity;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 
@@ -98,7 +101,7 @@ class AppTableTest extends TestCase
         Cache::write('testKey', 'testValue', $this->Posts->cache);
         $this->assertEquals('testValue', Cache::read('testKey', $this->Posts->cache));
 
-        $this->Posts->afterDelete(new \Cake\Event\Event(null), new \Cake\ORM\Entity, new \ArrayObject);
+        $this->Posts->afterDelete(new Event(null), new Entity, new ArrayObject);
 
         //The cache is cleared
         $this->assertFalse(Cache::read('testKey', $this->Posts->cache));
@@ -114,7 +117,7 @@ class AppTableTest extends TestCase
         Cache::write('testKey', 'testValue', $this->Posts->cache);
         $this->assertEquals('testValue', Cache::read('testKey', $this->Posts->cache));
 
-        $this->Posts->afterSave(new \Cake\Event\Event(null), new \Cake\ORM\Entity, new \ArrayObject);
+        $this->Posts->afterSave(new Event(null), new Entity, new ArrayObject);
 
         //The cache is cleared
         $this->assertFalse(Cache::read('testKey', $this->Posts->cache));